Create REST API endpoint for getting all customers
Anotate our Main class with @RestController in order to make it as a
web controller that cointains API endpoints for DispatherServlet use.
It will also serialize the returned objects into JSON syntax because
of @ResponseBody annotation contained in @RestController. Create method
getCustomers() and expose it as a REST API endpoint via @RequestMapping
annotation. Test it in RestfullTool plugin or via Postman.