Create custom unchecked exception ResourceNotFoundException
Replace the IllegalArgumentException in CustomerService with the new one.
Place this new exception in exception package under root package.
Anotate it with @ResponseStatus to get apropriate status code 404.
In application.yml, under server tab, include server error messages.
This is not a good practice because of possible PII leaks. Fix it later.